Move to_ methods to extract_util
